+++ src/main/java/webserver/ResponseHandler.java
@@ -3,6 +3,7 @@ package webserver;
 import com.github.jknack.handlebars.Template;
 import org.springframework.util.StringUtils;
 import utils.TemplateUtils;
+import webserver.http.Response;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
