+++ src/main/java/webserver/controller/Controller.java
@@ -1,6 +1,10 @@
-package webserver;
+package webserver.controller;
+
+import webserver.http.HttpRequest;
+import webserver.http.Response;
 
 @FunctionalInterface
+public
 interface Controller {
     Response execute(HttpRequest httpRequest) throws Exception;
