+++ src/main/java/webserver/controller/LoginController.java
@@ -1,7 +1,10 @@
-package webserver;
+package webserver.controller;
 
 import db.DataBase;
 import model.User;
+import webserver.http.HttpRequest;
+import webserver.http.Response;
+import webserver.http.ResponseStatus;
 
 import java.util.Map;
