+++ src/main/java/webserver/http/ResponseStatus.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 public enum ResponseStatus {
     OK("200 OK"),
